Added css selectable class Imported classnames module Added teamId to player details obj Removed old code

+++ nfl.fantasy.sanchez.bowl.webapp/nfl.fantasy.sanchez.bowl.domain/TeamIdentifier.cs
@@ -1,7 +1,7 @@
 ﻿namespace nfl.fantasy.sanchez.bowl.domain
 {
     //TODO: configure team names dyamically
-    public enum TeamIdentifier : uint
+    public enum TeamIdentifier : byte
     {
         NoTeam = 0,
         Vincent = 1,
